Extract TransientVector test to a separate class
