+++ y-collections/src/jvmTest/kotlin/com/github/whyrising/y/vector/PersistentVectorTest.kt
@@ -1078,298 +1078,6 @@ class PersistentVectorTest : FreeSpec({
         }
     }
 
-    "TransientVector" - {
-
-        @Suppress("UNCHECKED_CAST")
-        "constructor" {
-            val v = PersistentVector(*(1..57).toList().toTypedArray())
-            val vRoot = v.root
-
-            val tv = TransientVector(v)
-            val tvRoot = tv.root
-
-            tv.count shouldBeExactly v.count
-            tv.shift shouldBeExactly v.shift
-            tv.tail.size shouldBeExactly 32
-            assertArraysAreEquiv(tv.tail, v.tail)
-            tvRoot shouldNotBeSameInstanceAs vRoot
-            tvRoot.array.size shouldBeExactly vRoot.array.size
-            tvRoot.array shouldNotBeSameInstanceAs vRoot.array
-            tvRoot.isMutable shouldNotBeSameInstanceAs vRoot.isMutable
-            tvRoot.isMutable.value.shouldBeTrue()
-            vRoot.array.fold(0) { index: Int, e: Any? ->
-
-                if (e != null) {
-                    val tvNode = tvRoot.array[index] as Node<Int>
-                    val vNode = e as Node<Int>
-
-                    tvNode shouldBe vNode
-                    tvNode.isMutable shouldBeSameInstanceAs vRoot.isMutable
-                }
-                index + 1
-            }
-        }
-
-        "invalidate() should set isMutable to false" {
-            val v = PersistentVector(*(1..57).toList().toTypedArray())
-            val tv = TransientVector(v)
-            val isMutable = tv.root.isMutable
-
-            tv.invalidate()
-
-            tv.root.isMutable shouldBeSameInstanceAs isMutable
-            tv.root.isMutable.value.shouldBeFalse()
-        }
-
-        "assertMutable()" - {
-            "when called on a mutable transient, it shouldn't throw" {
-                val v = PersistentVector(*(1..57).toList().toTypedArray())
-                val tv = TransientVector(v)
-
-                shouldNotThrow<Exception> { tv.assertMutable() }
-            }
-
-            "when called on an invalidated transient, it should throw" {
-                val v = PersistentVector(*(1..57).toList().toTypedArray())
-                val tv = TransientVector(v)
-                tv.invalidate()
-
-                val e = shouldThrowExactly<IllegalStateException> {
-                    tv.assertMutable()
-                }
-
-                e.message shouldBe "Transient used after persistent() call"
-            }
-        }
-
-        "count" - {
-            """when called on a mutable transient,
-               it should return the count of the transient vector""" {
-                val v = PersistentVector(*(1..57).toList().toTypedArray())
-                val tv = TransientVector(v)
-
-                tv.root.isMutable.value.shouldBeTrue()
-                tv.count shouldBeExactly v.count
-            }
-
-            """when called on a invalidated transient,
-                                it should throw IllegalStateException""" {
-                val v = PersistentVector(*(1..57).toList().toTypedArray())
-                val tv = TransientVector(v)
-                tv.invalidate()
-
-                val e = shouldThrowExactly<IllegalStateException> { tv.count }
-
-                e.message shouldBe "Transient used after persistent() call"
-            }
-        }
-
-        "persistent()" - {
-            "when called on a invalidated transient, when it should throw" {
-                val v = PersistentVector(*(1..57).toList().toTypedArray())
-                val tv = TransientVector(v)
-                tv.invalidate()
-
-                val e = shouldThrowExactly<IllegalStateException> {
-                    tv.persistent()
-                }
-
-                e.message shouldBe "Transient used after persistent() call"
-            }
-
-            """when called on a mutable transient,
-                it should return the PersistentVector of that transient,
-                                                        and invalidate it""" {
-                val v = PersistentVector(*(1..57).toList().toTypedArray())
-                val tv = TransientVector(v)
-                tv.tail = v.tail.copyOf(32)
-
-                val vec = tv.persistent()
-                val root = vec.root
-
-                root.isMutable.value.shouldBeFalse()
-                root shouldBeSameInstanceAs tv.root
-                tv.root.isMutable.value.shouldBeFalse()
-                vec.count shouldBeExactly v.count
-                vec.shift shouldBeExactly v.shift
-                vec.tail.size shouldBeExactly 25
-                assertArraysAreEquiv(vec.tail, v.tail)
-            }
-        }
-
-        "conj(e:E)" - {
-
-            "when called on an invalidated transient, it should throw" {
-                val tempTv = TransientVector(PersistentVector(1, 2, 3))
-                tempTv.invalidate()
-
-                val e = shouldThrowExactly<IllegalStateException> {
-                    tempTv.conj(99)
-                }
-
-                e.message shouldBe "Transient used after persistent() call"
-            }
-
-            "when level is 5 and there is room in tail, it should add to it" {
-                val i = 7
-                val ints = Arb.int().filter { it != i }
-                checkAll(Arb.list(ints, 0..31)) { list ->
-                    val tempVec = PersistentVector(*list.toTypedArray())
-                    val tempTv = TransientVector(tempVec)
-
-                    val tv = tempTv.conj(i)
-                    val tvTail = tv.tail
-                    val tvRoot = tv.root
-
-                    tv.shift shouldBeExactly SHIFT
-                    tv.count shouldBeExactly list.size + 1
-                    tvRoot.isMutable.value.shouldBeTrue()
-                    tvRoot.array.size shouldBeExactly 32
-                    tvTail.size shouldBeExactly 32
-                    assertArraysAreEquiv(tvTail, tempVec.tail)
-
-                    tvTail[list.size] as Int shouldBeExactly i
-                }
-            }
-
-            "when tail overflow, it should push tail into the vec" - {
-                @Suppress("UNCHECKED_CAST")
-                "when level is 5, it should insert the tail in root node" {
-                    val listGen = Arb.list(Arb.int(), (32..1024)).filter {
-                        it.size % 32 == 0
-                    }
-
-                    checkAll(listGen, Arb.int()) { l: List<Int>, i: Int ->
-                        val tempVec = PersistentVector(*l.toTypedArray())
-                        val tempTv = TransientVector(tempVec)
-
-                        val tv = tempTv.conj(i)
-                        val tvTail = tv.tail
-                        val tvRoot = tv.root
-
-                        tv.shift shouldBeExactly SHIFT
-                        var index = 31
-                        var isMostRightLeafFound = false
-                        while (index >= 0 && !isMostRightLeafFound) {
-                            val node = tv.root.array[index]
-                            if (node != null) {
-                                val mostRightLeaf = node as Node<Int>
-                                val array = mostRightLeaf.array
-
-                                mostRightLeaf.isMutable shouldBeSameInstanceAs
-                                    tempTv.root.isMutable
-                                array shouldBe tempVec.tail
-                                mostRightLeaf.isMutable shouldBeSameInstanceAs
-                                    tempTv.root.isMutable
-                                isMostRightLeafFound = true
-                            }
-                            index--
-                        }
-
-                        tvRoot.isMutable.value.shouldBeTrue()
-                        isMostRightLeafFound.shouldBeTrue()
-                        tv.count shouldBeExactly l.size + 1
-                        tvTail.size shouldBeExactly 32
-                        tvTail[0] shouldBe i
-                    }
-                }
-
-                @Suppress("UNCHECKED_CAST")
-                """when level is > 5, it should iterate through the
-                                            levels then insert the tail""" {
-                    val e = 99
-                    val list = (1..1088).toList()
-                    val tempVec = PersistentVector(*list.toTypedArray())
-                    val root = tempVec.root
-                    val tempTv = TransientVector(tempVec)
-
-                    val tv = tempTv.conj(e)
-                    val tvTail = tv.tail
-                    val tvRoot = tv.root
-                    val tvSubRoot = tvRoot.array[1] as Node<Int>
-                    val firstLeft = tvSubRoot.array[0] as Node<Int>
-                    val mostRightLeaf = tvSubRoot.array[1] as Node<Int>
-
-                    tempTv.shift shouldBeExactly SHIFT * 2
-                    tvTail[0] as Int shouldBeExactly e
-                    assertArraysAreEquiv(mostRightLeaf.array, tempVec.tail)
-                    tv.count shouldBeExactly list.size + 1
-
-                    tvRoot.isMutable shouldNotBeSameInstanceAs root.isMutable
-                    tvRoot.isMutable.value.shouldBeTrue()
-                    tvSubRoot.isMutable shouldBeSameInstanceAs tvRoot.isMutable
-
-                    firstLeft.isMutable shouldBeSameInstanceAs root.isMutable
-                    mostRightLeaf.isMutable shouldBeSameInstanceAs
-                        tvRoot.isMutable
-                }
-
-                @Suppress("UNCHECKED_CAST")
-                """when level is > 5 and the subroot/path is null,
-                        it should create a new path then insert the tail""" {
-                    val e = 99
-                    val list = (1..2080).toList()
-                    val tempVec = PersistentVector(*list.toTypedArray())
-                    val tempTv = TransientVector(tempVec)
-
-                    val tv = tempTv.conj(e)
-                    val tvTail = tv.tail
-                    val tvRoot = tv.root
-                    val tvSubRoot = tvRoot.array[2] as Node<Int>
-                    val mostRightLeaf = tvSubRoot.array[0] as Node<Int>
-
-                    tv.shift shouldBeExactly SHIFT * 2
-                    assertArraysAreEquiv(mostRightLeaf.array, tempVec.tail)
-                    tvTail[0] shouldBe e
-                    tv.count shouldBeExactly list.size + 1
-                    tvRoot.isMutable.value.shouldBeTrue()
-                    tvSubRoot.isMutable shouldBeSameInstanceAs tvRoot.isMutable
-                    mostRightLeaf.isMutable shouldBeSameInstanceAs
-                        tvRoot.isMutable
-                }
-
-                @Suppress("UNCHECKED_CAST")
-                "when root overflow, it should add 1 lvl by creating new root" {
-                    val e = 99
-                    val list = (1..1056).toList()
-                    val tempVec = PersistentVector(*list.toTypedArray())
-                    val tempTv = TransientVector(tempVec)
-
-                    val tv = tempTv.conj(e)
-                    val tvRoot = tv.root
-                    val tvIsMutable = tvRoot.isMutable
-                    val subRoot1 = tvRoot.array[0] as Node<Int>
-                    val subRoot2 = tvRoot.array[1] as Node<Int>
-                    val mostRightLeaf =
-                        subRoot2.array[0] as Node<Int>
-
-                    tv.count shouldBeExactly list.size + 1
-                    tv.shift shouldBeExactly 10
-                    tv.tail[0] as Int shouldBeExactly e
-                    assertArraysAreEquiv(mostRightLeaf.array, tempVec.tail)
-
-                    tvIsMutable shouldBeSameInstanceAs tempTv.root.isMutable
-                    subRoot1.isMutable shouldBeSameInstanceAs tvIsMutable
-                    subRoot2.isMutable shouldBeSameInstanceAs tvIsMutable
-                    mostRightLeaf.isMutable shouldBeSameInstanceAs tvIsMutable
-                }
-            }
-        }
-
-        "concurrency" {
-            val v = v<Int>()
-            val transientVec = TransientVector(v)
-
-            withContext(Dispatchers.Default) {
-                massiveRun {
-                    transientVec.conj(1)
-                }
-            }
-
-            transientVec.count shouldBeExactly 100000
-        }
-    }
-
     "SubVector" - {
         "invoke()/ctor" - {
             "when start & end is out of bounds, it should throw exception" {
@@ -1690,7 +1398,7 @@ class PersistentVectorTest : FreeSpec({
     }
 }) {
     companion object {
-        private fun assertArraysAreEquiv(a1: Array<Any?>, a2: Array<Any?>) {
+        internal fun assertArraysAreEquiv(a1: Array<Any?>, a2: Array<Any?>) {
             a2.fold(0) { index: Int, i: Any? ->
                 val n = a1[index] as Int
 
@@ -1698,16 +1406,6 @@ class PersistentVectorTest : FreeSpec({
 
                 index + 1
             }
-        }
-    }
-}
-
-private suspend fun massiveRun(action: suspend () -> Unit) {
-    val n = 100
-    val times = 1000
-    coroutineScope {
-        repeat(n) {
-            launch { repeat(times) { action() } }
         }
     }
 }
